Miscellaneous fixes and cleanup (#24)
Fixes a number of small things including:

* Potential for self invocation of CopyGossipContactInfo to overwrite
real values
* Reading data offsets in CopyGossipContactInfo rather than hardcoding
assumed offsets
* Potential epoch overflow for cast epoch
* Comments around how specific structs and instructions are intended to
work
* Shifting epoch_start_timestamp to avoid overwriting existing bytes and
avoid trimming data type

+++ tests/tests/test_cluster_history.rs
@@ -113,10 +113,10 @@ async fn test_start_epoch_timestamp() {
 
     assert_eq!(account.history.arr[0].epoch, 0);
     assert_eq!(account.history.arr[1].epoch, 1);
-    assert_ne!(account.history.arr[0].epoch_start_timestamp, u32::MAX);
-    assert_ne!(account.history.arr[1].epoch_start_timestamp, u32::MAX);
+    assert_ne!(account.history.arr[0].epoch_start_timestamp, u64::MAX);
+    assert_ne!(account.history.arr[1].epoch_start_timestamp, u64::MAX);
     assert_eq!(
         account.history.arr[0].epoch_start_timestamp,
-        account.history.arr[1].epoch_start_timestamp - (dif_slots * MS_PER_SLOT) as u32
+        account.history.arr[1].epoch_start_timestamp - dif_slots * MS_PER_SLOT
     );
 }
